Move Subunits back to the symmetry.core package
It is not needed for clustering and very specific to quaternary symmetry
detection
