+++ biojava-structure-gui/src/main/java/org/biojava/nbio/structure/symmetry/jmolScript/JmolSymmetryScriptGeneratorH.java
@@ -23,8 +23,8 @@
  */
 package org.biojava.nbio.structure.symmetry.jmolScript;
 
-import org.biojava.nbio.structure.cluster.Subunits;
 import org.biojava.nbio.structure.symmetry.axis.HelixAxisAligner;
+import org.biojava.nbio.structure.symmetry.core.Subunits;
 import org.jcolorbrewer.ColorBrewer;
 
 import javax.vecmath.*;
