+++ biojava-structure/src/main/java/org/biojava/nbio/structure/cluster/SubunitCluster.java
@@ -36,6 +36,7 @@ import org.biojava.nbio.structure.align.fatcat.FatCatRigid;
 import org.biojava.nbio.structure.align.model.AFPChain;
 import org.biojava.nbio.structure.align.multiple.MultipleAlignment;
 import org.biojava.nbio.structure.align.multiple.MultipleAlignmentEnsembleImpl;
+import org.biojava.nbio.structure.symmetry.core.Subunits;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
