+++ biojava-structure/src/main/java/org/biojava/nbio/structure/symmetry/axis/AxisAligner.java
@@ -25,8 +25,8 @@ import javax.vecmath.Matrix4d;
 import javax.vecmath.Point3d;
 import javax.vecmath.Vector3d;
 
-import org.biojava.nbio.structure.cluster.Subunits;
 import org.biojava.nbio.structure.symmetry.core.QuatSymmetryResults;
+import org.biojava.nbio.structure.symmetry.core.Subunits;
 
 import java.util.List;
