+++ biojava-structure/src/main/java/org/biojava/nbio/structure/symmetry/core/HelixExtender.java
@@ -20,7 +20,6 @@
  */
 package org.biojava.nbio.structure.symmetry.core;
 
-import org.biojava.nbio.structure.cluster.Subunits;
 import org.biojava.nbio.structure.symmetry.geometry.SuperPosition;
 
 import javax.vecmath.Matrix4d;
