+++ biojava-structure/src/main/java/org/biojava/nbio/structure/symmetry/core/QuatSymmetryResults.java
@@ -22,7 +22,6 @@
  */
 package org.biojava.nbio.structure.symmetry.core;
 
-import org.biojava.nbio.structure.cluster.Subunits;
 
 /**
  * Holds the results of quaternary symmetry perception.
