+++ biojava-structure/src/test/java/org/biojava/nbio/structure/symmetry/core/TestQuatSymmetryDetection.java
@@ -71,6 +71,7 @@ public class TestQuatSymmetryDetection {
 		params.setParseBioAssembly(true);
 		params.setAlignSeqRes(true);
 		AtomCache cache = new AtomCache();
+		cache.setUseMmCif(true);
 		cache.setFileParsingParams(params);
 		StructureIO.setAtomCache(cache);
 		Structure pdb = StructureIO.getStructure("4hhb");
